\name{jobData}
\docType{data}
\alias{jobData}
\title{Sample Job Data for Clinical Psychologist}
\description{
  This data set contains job data for 'Clinical Psychologist'. It is the direct output of a \code{socSearch} using the O*NET SOC code \code{19-3031.02}, and is parsed into a list for efficient access by all package functions.
}
\usage{jobData}
\format{A list of length 16.}
\source{O*NET Online.}
\references{
  O*NET OnLine. \emph{National Center for O*NET Development}.
}
\keyword{datasets}