\name{jobData2}
\docType{data}
\alias{jobData2}
\title{Sample Job Data for Physical Therapist Aide}
\description{
  This data set contains job data for 'Physical Therapist Aide'. It is the direct output of a \code{socSearch} using the O*NET SOC code \code{31-2022.00}, and is parsed into a list for efficient access by all package functions.
}
\usage{jobData2}
\format{A list of length 16.}
\source{O*NET Online.}
\references{
  O*NET OnLine. \emph{National Center for O*NET Development}.
}
\keyword{datasets}